% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_weather.R
\name{.attach_weather}
\alias{.attach_weather}
\title{Join weather back to events (supports rolling join for hourly)}
\usage{
.attach_weather(
  x,
  weather,
  params,
  tz = "UTC",
  roll = c("nearest", "last", "none"),
  roll_max_hours = NULL,
  coord_digits = 5
)
}
\description{
Join weather back to events (supports rolling join for hourly)
}
\keyword{internal}
