% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_helpers.R
\name{.build_time}
\alias{.build_time}
\title{Build timestamp_raw from a single time column or multiple time-part columns}
\usage{
.build_time(x, time, tz = "UTC")
}
\description{
Build timestamp_raw from a single time column or multiple time-part columns
}
\keyword{internal}
