% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_plan.R
\name{.call_plan}
\alias{.call_plan}
\title{Plan provider calls: for each loc_id, split by time sparsity}
\usage{
.call_plan(
  x,
  time_col = "timestamp_utc",
  loc_id_col = "loc_id",
  rep_lat_col = "rep_lat",
  rep_lon_col = "rep_lon",
  tz = "UTC"
)
}
\description{
Plan provider calls: for each loc_id, split by time sparsity
}
\keyword{internal}
