% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{.elev_lookup}
\alias{.elev_lookup}
\title{Placeholder elevation lookup}
\usage{
.elev_lookup(lon, lat, method = c("constant"), constant = 100, ...)
}
\description{
Placeholder elevation lookup
}
\keyword{internal}
