% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_power.R
\name{.fetch_power}
\alias{.fetch_power}
\title{Fetch NASA POWER for planned calls}
\usage{
.fetch_power(
  calls_to_fetch,
  time_api,
  params,
  community = "ag",
  time_standard = "UTC",
  settings = list(),
  cache_dir = NULL,
  cache_scope = c("user", "project"),
  pkg = "weatherjoin",
  dummy_hour = 12L,
  verbose = FALSE,
  ...
)
}
\description{
Fetch NASA POWER for planned calls
}
\keyword{internal}
