% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_power_time.R
\name{.normalize_power_time}
\alias{.normalize_power_time}
\title{Normalize POWER output time columns to timestamp_utc (UTC)}
\usage{
.normalize_power_time(
  w,
  time_api = c("hourly", "daily"),
  tz = "UTC",
  dummy_hour = 12L
)
}
\description{
Normalize POWER output time columns to timestamp_utc (UTC)
}
\keyword{internal}
