% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_helpers.R
\name{.resolve_time_api}
\alias{.resolve_time_api}
\title{Resolve time_api based on user choice and input resolution}
\usage{
.resolve_time_api(
  dt,
  time_api = c("guess", "hourly", "daily"),
  input_res = c("hourly", "daily"),
  tz = "UTC",
  dummy_hour = 12L
)
}
\description{
Resolve time_api based on user choice and input resolution
}
\keyword{internal}
