% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plan.R
\name{.spatial_plan}
\alias{.spatial_plan}
\title{Spatial planning: map points to representative locations}
\usage{
.spatial_plan(
  x,
  spatial_mode = c("cluster", "exact", "by_group"),
  lat_col = "lat",
  lon_col = "lon",
  group_col = NULL,
  rep_method = c("median", "centroid"),
  cluster_radius_m = 250,
  keep_diag = TRUE,
  check_range = TRUE,
  coord_digits = 5L
)
}
\description{
Spatial planning: map points to representative locations
}
\keyword{internal}
