% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_time.R
\name{.split_time_ranges}
\alias{.split_time_ranges}
\title{Split sparse time points into segments using a gap penalty (hours)}
\usage{
.split_time_ranges(
  times_utc,
  split_penalty_hours = 72,
  pad_hours = 0,
  max_parts = 50L
)
}
\description{
Split sparse time points into segments using a gap penalty (hours)
}
\keyword{internal}
