% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wj_load}
\alias{.wj_load}
\title{Internal: load required packages (used for interactive sourcing too)}
\usage{
.wj_load(pkgs = c("data.table"), attach = FALSE, quiet = TRUE)
}
\description{
Internal: load required packages (used for interactive sourcing too)
}
\keyword{internal}
