% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_imports.R
\docType{package}
\name{weatherjoin-package}
\alias{weatherjoin}
\alias{weatherjoin-package}
\title{weatherjoin: Join Gridded Weather Data to Event Tables}
\description{
High-level tools to attach gridded weather data from the NASA POWER project to event-based datasets. The package plans efficient spatio-temporal API calls, caches downloaded segments locally, and joins weather variables back to the input table using exact or rolling joins. NASA POWER data are retrieved via the 'nasapower' R package. This package is not affiliated with or endorsed by NASA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hauae/weatherjoin}
  \item Report bugs at \url{https://github.com/hauae/weatherjoin/issues}
}

}
\author{
\strong{Maintainer}: Przemek Dolowy \email{pdolowy@harper-adams.ac.uk} (Harper Adams University)

}
\keyword{internal}
